+++ pyspedas/projects/mms/databar_tools/mms_load_sroi_segments.py
